+++ docker/dev/test_deploy.sh
@@ -4,7 +4,9 @@ set -e
 BRANCH=$1
 
 git checkout $BRANCH
-cd ../../
+
+export COMPOSE_DOCKER_CLI_BUILD=1
+export DOCKER_BUILDKIT=1
 
 # Rebuilding dev docker
 sudo docker build . -f "docker/dev/Dockerfile" -t "dessalines/lemmy:$BRANCH"