+++ docker/federation/run-federation-test.bash
@@ -20,4 +20,9 @@ popd || exit
 
 sudo docker build ../../ --file Dockerfile -t lemmy-federation:latest
 
+for Item in alpha beta gamma ; do
+  sudo mkdir -p volumes/pictrs_$Item
+  sudo chown -R 991:991 volumes/pictrs_$Item
+done
+
 sudo docker-compose up