+++ docker/prod/deploy.sh
@@ -1,4 +1,5 @@
 #!/bin/sh
+set -e
 git checkout master
 
 # Import translations
@@ -20,7 +21,7 @@ git add "server/src/version.rs"
 echo $new_tag > "ansible/VERSION"
 git add "ansible/VERSION"
 
-cd docker/dev || exit
+cd docker/prod || exit
 
 # Changing the docker-compose prod
 sed -i "s/dessalines\/lemmy:.*/dessalines\/lemmy:$new_tag/" ../prod/docker-compose.yml
@@ -32,40 +33,23 @@ git add ../../ansible/templates/docker-compose.yml
 git commit -m"Version $new_tag"
 git tag $new_tag
 
+export COMPOSE_DOCKER_CLI_BUILD=1
+export DOCKER_BUILDKIT=1
+
 # Rebuilding docker
-docker-compose build
-docker tag dev_lemmy:latest dessalines/lemmy:x64-$new_tag
-docker push dessalines/lemmy:x64-$new_tag
-
-# Build for Raspberry Pi / other archs
-
-# Arm currently not working
-# docker build -t lemmy:armv7hf -f Dockerfile.armv7hf ../../
-# docker tag lemmy:armv7hf dessalines/lemmy:armv7hf-$new_tag
-# docker push dessalines/lemmy:armv7hf-$new_tag
-
-# aarch64
-# Only do this on major releases (IE the third semver is 0)
 if [ $third_semver -eq 0 ]; then
-  # Registering qemu binaries
-  docker run --rm --privileged multiarch/qemu-user-static:register --reset
-
-  docker build -t lemmy:aarch64 -f Dockerfile.aarch64 ../../
-  docker tag lemmy:aarch64 dessalines/lemmy:arm64-$new_tag
-  docker push dessalines/lemmy:arm64-$new_tag
-fi
-
-# Creating the manifest for the multi-arch build
-if [ $third_semver -eq 0 ]; then
-  docker manifest create dessalines/lemmy:$new_tag \
-  dessalines/lemmy:x64-$new_tag \
-  dessalines/lemmy:arm64-$new_tag
+  # TODO get linux/arm/v7 build working
+  # Build for Raspberry Pi / other archs too
+  docker buildx build --platform linux/amd64,linux/arm64 ../../ \
+    --file Dockerfile \
+    --tag dessalines/lemmy:$new_tag \
+    --push
 else
-  docker manifest create dessalines/lemmy:$new_tag \
-  dessalines/lemmy:x64-$new_tag
+  docker buildx build --platform linux/amd64 ../../ \
+    --file Dockerfile \
+    --tag dessalines/lemmy:$new_tag \
+    --push
 fi
-
-docker manifest push dessalines/lemmy:$new_tag
 
 # Push
 git push origin $new_tag